LayoutTests/platform/chromium/media/video-black-bg-in-media-document-expected.txt

Test that video media documents have a black background.


EXPECTED (iframeBodyStyle.backgroundColor == 'rgb(0, 0, 0)') OK
END OF TEST
